@@ -38,7 +38,7 @@ public class TestPizzaV1Core : QueryTestBase
 
         if (resultCreate.IsError)
         {
-            Assert.IsTrue(false);
+            Assert.That(false, Is.False);
         }
 
         this.model = resultCreate.Data;
@@ -54,7 +54,7 @@ public class TestPizzaV1Core : QueryTestBase
                 Id = this.model.Id
             }, CancellationToken.None);
 
-        Assert.IsTrue(resultGet?.Data != null);
+        Assert.That(resultGet?.Data, Is.Not.Null);
     }
 
     [Test]
@@ -70,11 +70,12 @@ public class TestPizzaV1Core : QueryTestBase
                 }
             }, CancellationToken.None);
 
-        Assert.IsTrue(resultGetAll?.Data.Count > 0);
+        Assert.That(resultGetAll?.Data.Count > 0, Is.True);
     }
 
     [Test]
-    public void SaveAsync() => Assert.IsTrue(this.model != null);
+    public void SaveAsync()
+        => Assert.That(this.model, Is.Not.Null);
 
     [Test]
     public async Task UpdateAsync()
@@ -90,7 +91,7 @@ public class TestPizzaV1Core : QueryTestBase
                 }
             }, CancellationToken.None);
 
-        Assert.IsTrue(!resultUpdate.IsError);
+        Assert.That(resultUpdate.IsError, Is.False);
     }
 
     [Test]
@@ -103,6 +104,6 @@ public class TestPizzaV1Core : QueryTestBase
                 Id = this.model.Id
             }, CancellationToken.None);
 
-        Assert.IsTrue(!outcomeDelete.IsError);
+        Assert.That(outcomeDelete.IsError, Is.False);
     }
 }
