@@ -26,7 +26,7 @@ public class TestPizzaeV1DataAccess : QueryTestBase
         var result = await this.handler.Save(PizzaTestData.Create);
         if (result.IsError)
         {
-            Assert.IsTrue(false);
+            Assert.That(false, Is.False);
         }
 
         this.model = result.Data;
@@ -36,7 +36,7 @@ public class TestPizzaeV1DataAccess : QueryTestBase
     public async Task GetAsync()
     {
         var response = await this.handler.Get(this.model.Id);
-        Assert.IsTrue(response != null);
+        Assert.That(response, Is.Not.Null);
     }
 
     [Test]
@@ -48,25 +48,24 @@ public class TestPizzaeV1DataAccess : QueryTestBase
         });
         var outcome = response.Count;
 
-        Assert.IsTrue(outcome == 1);
+        Assert.That(outcome == 1, Is.True);
     }
 
     [Test]
     public async Task UpdateAsync()
     {
-        var originalModel = this.model;
         this.model.Name = "Test";
         var response = await this.handler.Update(this.model.Id, PizzaTestData.Update);
-        Assert.IsTrue(response != null && !response.IsError);
+        Assert.That(response != null && !response.IsError, Is.True);
 
         var outcome = response.Data.Name.Equals(PizzaTestData.Update.Name);
-        Assert.IsTrue(outcome);
+        Assert.That(outcome, Is.True);
     }
 
     [Test]
     public async Task DeleteAsync()
     {
         var response = await this.handler.Delete(this.model.Id);
-        Assert.IsTrue(response != null && !response.IsError);
+        Assert.That(response != null && !response.IsError, Is.True);
     }
 }
